Fix spring boot 2.6 compatibility
https://issues.redhat.com/browse/ENT-4649

Essentially, it looks like changes to how deps are resolved in Spring
Boot applications made it so that Spring was unable to resolve the beans
for RBAC in our security config. Moving them to a separate configuration
class and then simply depending on that via `@Import` has resolved the
issue.
